afdafdafasfaf

+++ Carnivore.cpp
@@ -3,9 +3,10 @@
 using namespace std;
 
 #define factor 0.5
+
 int Carnivore :: nCarnivore = 0;
 int Carnivore::  total_cFood = 0;
-Carnivore :: Carnivore(int weight) : cFood(weight * factor) {
+Carnivore :: Carnivore(float weight) : cFood(weight * factor) {
 	nCarnivore++;
 	total_cFood += cFood;
 }
