+++ Carnivore.h
@@ -9,11 +9,11 @@ public:
 	/** @brief Constructor.
 	  * Membuat objek hewan karnivora dan mengincrement atribut nCarnivore
 	  */
-	Carnivore(int weight);
+	Carnivore(float weight);
 	static int nCarnivore; 	/** Jumlah hewan karnivora */
 	static int total_cFood;
 protected:
-	int cFood;
+	float cFood;
 };
 
 #endif
